src/requirements.txt
rdflib
flask
jinja2
gitpython